Moved platform implementations into their own modules
